update to operate on cluster
